+++ packages/api/lib/app.ts
@@ -1,6 +1,7 @@
 import { oc } from "@orpc/contract";
 import { z } from "zod";
 
+import { AccountManagerAPIContract } from "./account";
 import { AppletManagerAPIContract } from "./applet";
 import { AuthAPIContract } from "./auth";
 import { SessionManagerAPIContract } from "./session";
@@ -12,6 +13,7 @@ export const AppAPIContract = {
             version: z.literal(0),
         })
     ),
+    accountManager: AccountManagerAPIContract,
     appletManager: AppletManagerAPIContract,
     auth: AuthAPIContract,
     sessionManager: SessionManagerAPIContract,