+++ packages/serverkit-applet-vscode/src/index.ts
@@ -7,7 +7,7 @@ export default ((registry) => {
     registry.applet.spawners.set(
         "vscode",
         async () => {
-            return VSCodeAppletSpawner();
+            return await VSCodeAppletSpawner();
         },
     );
 }) as IKit;