+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hagateway/server
 
+## 0.0.1-alpha.7
+
+### Patch Changes
+
+- Internal test release.
+- Updated dependencies
+  - @hagateway/api@0.0.1-alpha.5
+
 ## 0.0.1-alpha.6
 
 ### Patch Changes